subscription fixes 1.3

+++ backend/foodgram/users/views.py
@@ -1,22 +1,20 @@
 """Представления приложения users."""
 
 from django.shortcuts import get_object_or_404
-from rest_framework import status, views, filters
+from rest_framework import status, views
 from rest_framework.generics import ListAPIView
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
 from core.pagination import CustomPagination
 from users.models import Subscription, User
-from .serializers import SubscribeSerializer, SubscriptionSerializer
+from .serializers import SubscriptionSerializer
 
 
 class SubscriptionViewSet(ListAPIView):
     serializer_class = SubscriptionSerializer
     pagination_class = CustomPagination
     permission_classes = (IsAuthenticated,)
-    filter_backends = (filters.SearchFilter,)
-    search_fields = ('^subscribed__user',)
 
     def get_queryset(self):
         user = self.request.user
@@ -26,23 +24,44 @@ class SubscriptionViewSet(ListAPIView):
 class SubscribeView(views.APIView):
     pagination_class = CustomPagination
     permission_classes = (IsAuthenticated,)
+    serializer_class = SubscriptionSerializer
 
-    def post(self, request, pk):
-        author = get_object_or_404(User, pk=pk)
-        user = self.request.user
-        data = {'author': author.id, 'user': user.id}
-        serializer = SubscribeSerializer(
-            data=data, context={'request': request}
+    def post(self, request, *args, **kwargs):
+        user_id = self.kwargs.get('user_id')
+        if user_id == request.user.id:
+            return Response(
+                {'error': 'Нельзя подписаться на себя'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        if Subscription.objects.filter(
+                user=request.user,
+                author_id=user_id
+        ).exists():
+            return Response(
+                {'error': 'Вы уже подписаны на пользователя'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        author = get_object_or_404(User, id=user_id)
+        Subscription.objects.create(
+            user=request.user,
+            author_id=user_id
         )
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(data=serializer.data, status=status.HTTP_201_CREATED)
+        return Response(
+            self.serializer_class(author, context={'request': request}).data,
+            status=status.HTTP_201_CREATED
+        )
 
-    def delete(self, request, pk):
-        author = get_object_or_404(User, pk=pk)
-        user = self.request.user
-        subscription = get_object_or_404(
-            Subscription, user=user, author=author
+    def delete(self, request, *args, **kwargs):
+        user_id = self.kwargs.get('user_id')
+        get_object_or_404(User, id=user_id)
+        subscription = Subscription.objects.filter(
+            user=request.user,
+            author_id=user_id
         )
-        subscription.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if subscription:
+            subscription.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(
+            {'error': 'Вы не подписаны на пользователя'},
+            status=status.HTTP_400_BAD_REQUEST
+        )
